修改group的定义，将其作为联赛或者杯赛的字段

+++ src/main/java/de/fred4jupiter/fredbet/domain/AppUser.java
@@ -61,7 +61,7 @@ public class AppUser implements UserDetails {
     private boolean firstLogin;
     
     @Column(name = "POINTS")
-    private int points;
+    private Double points;
 
     @PersistenceCreator
     protected AppUser() {
@@ -250,11 +250,11 @@ public class AppUser implements UserDetails {
         this.firstLogin = firstLogin;
     }
 
-	public int getPoints() {
+	public Double getPoints() {
 		return points;
 	}
 
-	public void setPoints(int points) {
+	public void setPoints(Double points) {
 		this.points = points;
 	}
