+++ src/main/java/de/fred4jupiter/fredbet/service/user/UserService.java
@@ -104,12 +104,12 @@ public class UserService {
     }
 
     @CacheEvict(cacheNames = CacheNames.CHILD_RELATION, allEntries = true)
-    public AppUser updateUser(Long userId, boolean isChild, int points) {
+    public AppUser updateUser(Long userId, boolean isChild, Double points) {
         return updateUser(userId, null, isChild, points);
     }
 
     @CacheEvict(cacheNames = CacheNames.CHILD_RELATION, allEntries = true)
-    public AppUser updateUser(Long userId, Set<String> roles, boolean isChild, int points) {
+    public AppUser updateUser(Long userId, Set<String> roles, boolean isChild, Double points) {
         Assert.notNull(userId, "userId must be given");
         AppUser appUser = findByUserId(userId);
         if (Validator.isNotEmpty(roles)) {
