+++ src/main/java/de/fred4jupiter/fredbet/web/info/PointsFrequencyCommand.java
@@ -13,11 +13,11 @@ public class PointsFrequencyCommand {
 
 	private final List<PointsFrequency> resultList = new ArrayList<>();
 
-	public PointsFrequencyCommand(MultiValuedMap<Integer, PointCountResult> map) {
-		List<Integer> pointsList = new ArrayList<>(map.keySet());
+	public PointsFrequencyCommand(MultiValuedMap<Double, PointCountResult> map) {
+		List<Double> pointsList = new ArrayList<>(map.keySet());
 		Collections.reverse(pointsList);
 
-		for (Integer points : pointsList) {
+		for (Double points : pointsList) {
 			if (points > 0) {
 				Collection<PointCountResult> collection = map.get(points);
 				PointsFrequency pointsFrequency = new PointsFrequency(points, new ArrayList<>(collection));
